data-analysis/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
futures==3.1.1
gitdb==4.0.6
GitPython==3.1.14
greenlet==1.0.0
idna==2.10
ijson==3.1.4
kiwisolver==1.3.1
lizard==1.17.7
matplotlib==3.4.1
mpmath==1.2.1
mysql-connector-python==8.0.25
natsort==7.1.1
numpy==1.20.2
pandas==1.2.3
patsy==0.5.1
Pillow==8.1.2
powerlaw==1.4.6
protobuf==3.15.8
PyDriller==1.15.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
RequestsThrottler==0.2.5
scipy==1.6.2
seaborn==0.11.1
semantic-version==2.8.5
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.4.9
statsmodels==0.12.2
tld==0.12.5
tqdm==4.60.0
urllib3==1.26.4

python-dotenv~=0.19.2
rbo~=0.1.2